real-voice-assistant/requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.11
pyttsx3>=2.90
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiofiles>=23.0.0
asyncio>=3.4.3
rich>=13.0.0
elevenlabs>=1.1.0
openai-whisper>=202311
